refactor: 🧹 Jules05: Product coherence improvements
- Relocate FeatureEngineer from src/core to src/data for better domain separation.
- Consolidate RiskEngine into RiskManager and implement a harmonized 8-layer ATR-based risk cascade.
- Standardize Risk Management API on `validate_signal()` returning `RiskDecision`.
- Update `AuditedRiskManager` with full decision chain audit logging.
- Align `main.py` and documentation (README.md) with the new institutional standards.
- Cleanup redundant modules and redundant exports.
